boletasConsorcios/requirements.txt
django-easy-pdf==0.1.0
django-widget-tweaks==1.4.1
fpdf==1.7.2
html5lib==0.999999
Pillow==2.9.0
pycparser==2.14
PyPDF2==1.25
Pyphen==0.9.1
python-dateutil==2.4.2
six==1.9.0
sqlparse==0.1.15
tinycss==0.3
wheel==0.24.0
xhtml2pdf==0.0.6
